requirements.txt
asyncpg==0.27.0
discord.py[voice, speed]==2.3.1
python-dotenv
wavelink==2.5.1
